Get rid of bad redundant HZB call

+++ D3D12/Graphics/ShaderInterop.h
@@ -183,7 +183,7 @@ namespace ShaderInterop
 		float2 ViewportDimensions;
 		float2 ViewportDimensionsInv;
 		float2 ViewJitter;
-		float2 HZBDimensions;
+		int2 HZBDimensions;
 		float NearZ;
 		float FarZ;
 		float FoV;
